Refactor utilities folder in useCases to be able to test the functions without passing requests/responses as function parameter
