+++ README.md
@@ -10,3 +10,4 @@
 - Error Handler Middleware
 - Logging
 - Content of Landing Page
+- Rename getAudioFileById to get getFile and change all definitions (e. g. route path) accordingly
